Contest improvements:
- Support for interactive problem type
- Option to allow multiple correct submissions for a problem
- Ability to make one problem depend on others (so that they must be solved
  first)
- "Format Error" judge option, in addition to old "Correct" and "Wrong"
- Python support
- Allow server config to specify supported languages

UI improvements:
- New-look Javascript-based scores page
- Allow a team name separately from a login ID
- Clarification request option on submissions tab
- Notification icons on tabs to indicate new information
- Standings tab no longer has leading zeros on positions.

Admin/judge improvements:
- Stderr from a run is included in judges' view
- New command-line tools to
  - add problems
  - extract the latest submission for a problem for a user
  - submit a problem
  - submit generic commands (including file uploads), mostly for testing
- Now possible to specify only one of start and stop time.
- Initial password can now be set in config file

Network improvements:
- Marking eventually times out, so that a wedged marker does not hold
  everything up.
- Marker now attempts to auto-reconnect if the server goes down
- Submissions are now capped at 100KB.
- Server pushes necessary information to the client rather than the client
  pulling all data synchronously. This gives massive improvements in
  interactivity.
- Keepalive mechanism, to avoid problems with ISPs that kill off idle
  connections.
- Introduce ThreadSSL class to provide safer user of OpenSSL

Other:
- Even more verbose logging in the server.
- A suite of test programs that do evil things.
- No longer requires boost.
- Observer role changed slightly.

Bug fixes:
- searches for "public class Foo" when trying to detect Java classname
- java.policy file now actually takes effect
- OTHER state should not be counted as a wrong submission
- allow System.setIn and System.setOut to be used in Java
- Unicode now works in many more places
- Forces LC_ALL=C in marker to give consistent results across markers
- Clean up temporary files left by marker
- Tidy up some error messages to avoid duplicate errors and to provide more
  helpful hints.
- A database leak fixed.
- Some locking errors fixed.
- Use _exit instead of exit when quitting from a failed fork()/exec()
- Fix assorted compilation warnings
- Allow multiple connections from a single client to work properly.
- Remove some UI controls for which there was no code
- Fix some bugs around registering and deregistring clients that could corrupt
  the registry
- Fixes for lots of race conditions and memory leaks
- Fix race conditions in problem ID generation


git-svn-id: https://abacuscm.svn.sourceforge.net/svnroot/abacuscm/trunk@728 ac3b2d11-31fc-0310-b7bc-aecf6fe4fc4f

+++ src/servereventregistry.C
@@ -44,7 +44,7 @@ ServerEventRegistry::~ServerEventRegistry()
 
 	for (i = _eventlist.begin(); i != _eventlist.end(); ++i) {
 		if (i->second.listeners.size())
-			log(LOG_WARNING, "Event '%s' has %lu registered subscribers at shutdown left.", i->first.c_str(), i->second.listeners.size());
+			log(LOG_WARNING, "Event '%s' has %lu registered subscribers at shutdown left.", i->first.c_str(), (unsigned long) i->second.listeners.size());
 		pthread_mutex_destroy(&i->second.lock);
 	}
